feat(ts-jest): add test-jest
