+++ .gitignore
@@ -1,3 +1,9 @@
 # yarn.lock
 node_modules
 dist
+
+# vscode
+.vscode
+
+# jest coverage
+coverage
