+++ README.md
@@ -1,7 +1,15 @@
 # JSBD
 
-## a JSBI style polyfill
+## a JSBI style polyfill for BigDecimal
 
-because use es module in node , better use node >= 14
+base on native BigInt implement
 
-try to make husky work
+## for develop
+
+since test use es module in node , better use node >= 14
+
+```
+yarn install
+// do something (and add test case)
+yarn test
+```
