+++ jest.config.js
@@ -1,1 +1,5 @@
-export default {}
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+}
