--- src/bigint/type.ts
@@ -1,10 +1,0 @@
-export type DecimalIntVal = string | number
-
-export type DecimalSign = -1 | 1
-
-export type Exponent = { sign: DecimalSign; value: string }
-
-export interface RoundOption {
-  maximumFractionDigits?: number
-  roundingMode?: 'down' | 'half down' | 'half up' | 'half even' | 'up'
-}
