+++ src/decimal.ts
@@ -1,22 +1,23 @@
-import { DecimalIntVal, DecimalSign, Exponent } from './type'
 import { isLiteral, parseLiteral } from './utils'
 
 export class Decimal {
-  sign: DecimalSign
-  mantissa: string
-  exponent: Exponent
-  constructor(intVal: DecimalIntVal) {
+  mantissa!: bigint
+  exponent!: number
+  constructor(intVal: number | string | bigint) {
     switch (typeof intVal) {
-      // case 'number': {
-      //   // this.mantissa = intVal.toString(10)
-      //   break
-      // }
+      case 'bigint': {
+        return new Decimal(intVal.toString(10))
+      }
+      case 'number': {
+        return new Decimal(intVal.toString(10))
+      }
       case 'string': {
         if (isLiteral(intVal)) {
-          let { sign, mantissa, exponent } = parseLiteral(intVal)
-          this.sign = sign
+          let { mantissa, exponent } = parseLiteral(intVal)
           this.mantissa = mantissa
           this.exponent = exponent
+        } else if (isLiteral(intVal)) {
+          // to do scientific notation
         } else {
           throw new Error('string value must be legal number')
         }
@@ -24,6 +25,37 @@ export class Decimal {
       }
       default: {
         throw new Error('init value must be string or number')
+      }
+    }
+  }
+  toString() {
+    if (this.mantissa === 0n) {
+      return '0'
+    }
+    if (this.exponent >= 0) {
+      let zero = ''
+      let offset = this.exponent
+      while (offset > 0) {
+        zero += '0'
+        offset--
+      }
+      return this.mantissa.toString(10) + zero
+    } else {
+      let offset = this.exponent * -1
+      let str = this.mantissa.toString(10)
+      if (str.length > offset) {
+        let minus = str.length - offset
+        let withZeroStr = str.slice(0, minus) + '.' + str.slice(minus)
+        return withZeroStr.replace(/\.?0*$/, '')
+      } else {
+        let minus = offset - str.length
+        let addZero = '0.'
+        while (minus > 0) {
+          addZero += '0'
+          minus--
+        }
+        let withZeroStr = addZero + str
+        return withZeroStr.replace(/\.?0*$/, '')
       }
     }
   }
