--- test/equal.test.js
@@ -1,7 +1,0 @@
-import { JSBD } from '../dist/jsbd.js'
-
-test('"100" isEqual "100.000"', () => {
-  // let a = JSBD.BigDecimal('100')
-  // let b = JSBD.BigDecimal('100.000')
-  // expect(JSBD.isEqual(a, b)).toBeTruthy()
-})
