+++ tsconfig.json
@@ -7,7 +7,8 @@
     "noImplicitAny": true,
     "sourceMap": false,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "esModuleInterop": true
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"]
